Change beat resolution to 120 ticks.
This permits perfect triplets, quintuplets, sextuplets, and 32nd notes,
if a beat is a quarter note. I wanted to add septuplets, but they're so rare
and they turn the resolution *way* up, and they can be done with 17 ticks each
and an error of 1 tick per beat. (And let's be honest, nobody can hear 120th
notes.)